requirements.txt
streamlit
pymupdf
torch
transformers
sentence-transformers
faiss-cpu
pdfplumber
Pillow
